added get players method
